# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Color API
  description: API pour gérer les stocks des couleurs.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Entrez votre token JWT (obtenu via /login)"
...
paths:
  /login:
    post:
      tags:
        - User
      summary: Authentifier un utilisateur et obtenir un token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "password123"
              required:
                - username
                - password
      responses:
        '200':
          description: Authentification réussie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Requête invalide (données manquantes ou mal formées).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid request data"
        '401':
          description: Échec de l'authentification (nom d'utilisateur ou mot de passe incorrect).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid username or password"
        '500':
          description: Erreur serveur.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Server error"
...
  /users:
    get:
      tags:
        - Admin
      summary: Obtenir la liste de tous les utilisateurs.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Récupère la liste de tous les utilisateurs non supprimés.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des utilisateurs récupérée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Users retrieved successfully"
                  valid:
                    type: boolean
                    example: true
                  value:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        description:
                          type: string
                          description: "Description de l'utilisateur"
                          example: "Administrateur principal"
                        username:
                          type: string
                          example: "admin"
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-01-01T10:00:00Z"
                        last_connexion:
                          type: string
                          format: date-time
                          example: "2024-10-06T14:30:00Z"
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '500':
          description: Erreur serveur.
  /addUser:
    post:
      tags:
        - Admin
      summary: Créer un nouvel utilisateur.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Crée un nouveau compte utilisateur avec les informations fournies.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  description: "Nom d'utilisateur unique"
                  example: "newuser"
                description:
                  type: string
                  description: "Description de l'utilisateur"
                  example: "Nouvel utilisateur"
                password:
                  type: string
                  minLength: 6
                  description: "Mot de passe (minimum 6 caractères)"
                  example: "securepassword123"
              required:
                - username
                - password
      responses:
        '201':
          description: Utilisateur créé avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created successfully"
                  status:
                    type: integer
                    example: 201
                  valid:
                    type: boolean
                    example: true
                  value:
                    type: object
                    description: "Données de l'utilisateur créé"
        '400':
          description: Données invalides ou utilisateur déjà existant.
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '500':
          description: Erreur serveur.
  /editUser/{id}:
    post:
      tags:
        - Admin
      summary: Modifier les données d'un utilisateur.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Modifie les informations et permissions d'un utilisateur existant.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: "ID unique de l'utilisateur à modifier"
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  description: "Nouveau nom d'utilisateur"
                  example: "updateduser"
                description:
                  type: string
                  description: "Nouvelle description de l'utilisateur"
                  example: "Utilisateur modifié"
                permissions:
                  type: array
                  items:
                    type: integer
                  description: "Liste des IDs des permissions à attribuer"
                  example: [1, 2, 3]
      responses:
        '200':
          description: Utilisateur modifié avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated successfully"
                  status:
                    type: integer
                    example: 200
                  valid:
                    type: boolean
                    example: true
        '400':
          description: Données invalides ou ID non valide.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "id is not valid"
                  status:
                    type: integer
                    example: 400
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur serveur.
  /deleteUser/{id}:
    post:
      tags:
        - Admin
      summary: Supprimer un utilisateur (suppression logique).
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Effectue une suppression logique de l'utilisateur (marque comme supprimé).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: "ID unique de l'utilisateur à supprimer"
          example: 1
      responses:
        '200':
          description: Utilisateur supprimé avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"
                  status:
                    type: integer
                    example: 200
                  valid:
                    type: boolean
                    example: true
        '400':
          description: ID invalide (doit être un entier positif).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "id is not valid"
                  status:
                    type: integer
                    example: 400
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur serveur.
  /permissions:
    get:
      tags:
        - Admin
      summary: Obtenir la liste de toutes les permissions disponibles.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Récupère la liste complète des permissions disponibles dans le système.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des permissions récupérée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Permissions retrieved successfully"
                  valid:
                    type: boolean
                    example: true
                  value:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: "admin"
                        description:
                          type: string
                          example: "Accès complet au système"
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '500':
          description: Erreur serveur.
  /permissions/{id}:
    get:
      tags:
        - Admin
      summary: Obtenir les permissions d'un utilisateur spécifique.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Récupère la liste des permissions attribuées à un utilisateur donné.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: "ID unique de l'utilisateur"
          example: 1
      responses:
        '200':
          description: Permissions de l'utilisateur récupérées avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User permissions retrieved successfully"
                  value:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: "visitor"
                        description:
                          type: string
                          example: "Accès en lecture seule"
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur serveur.
  /logs:
    get:
      tags:
        - Admin
      summary: Obtenir les logs de l'application.
      description: |
        ⚠️ **Permissions requises** : admin uniquement
        
        Récupère les logs d'activité et d'erreur de l'application pour le monitoring.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logs récupérés avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logs retrieved successfully"
                  value:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        action:
                          type: string
                          example: "COLOR_ADDED"
                        user_id:
                          type: integer
                          example: 1
                        timestamp:
                          type: string
                          format: date-time
                          example: "2024-10-06T14:30:00Z"
                        details:
                          type: string
                          example: "Nouvelle couleur RAL3000 ajoutée"
        '401':
          description: Token manquant ou invalide.
        '403':
          description: Permissions insuffisantes (admin requis).
        '500':
          description: Erreur serveur.